Fix DBCP-373 Modify SharedPoolDataSource and PerUserPoolDataSource to expose all of the configuration properties of the underlying connection pool(s).
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/dbcp/trunk@1568055 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/java/org/apache/commons/dbcp2/datasources/TestCPDSConnectionFactory.java
@@ -66,7 +66,7 @@ public class TestCPDSConnectionFactory extends TestCase {
        PerUserPoolDataSource ds = new PerUserPoolDataSource();
        ds.setConnectionPoolDataSource(cpds);
        ds.setPerUserMaxTotal("username", Integer.valueOf(10));
-       ds.setPerUserMaxWaitMillis("username", Integer.valueOf(50));
+       ds.setPerUserMaxWaitMillis("username", Long.valueOf(50));
        ds.setPerUserMaxIdle("username", Integer.valueOf(2));
        Connection conn1 = ds.getConnection("username", "password");
        Connection conn2 = ds.getConnection("username", "password");
